+++ Components/Schumix.Components/LuaEngine/SchumixCommandMethod.cs
@@ -19,22 +19,22 @@
  */
 
 using System;
-using System.IO;
-using System.Net;
-using NGit;
-using NGit.Api;
+using Schumix.Api.Delegate;
+using Schumix.Irc.Commands;
 
-namespace Schumix.Updater.Download
+namespace Schumix.Components.LuaEngine
 {
-	sealed class CloneSchumix
+	#pragma warning disable 3015
+	public class SchumixCommandMethod : Attribute
 	{
-		public CloneSchumix(string Url, string Dir, string Branch)
+		public CommandPermission Permission { get; private set; }
+		public CommandDelegate Method { get; set; }
+
+		public SchumixCommandMethod(CommandDelegate method, CommandPermission permission = CommandPermission.Normal)
 		{
-			var clone = Git.CloneRepository();
-			var repo = clone.SetURI(Url).SetDirectory(Path.Combine(Environment.CurrentDirectory, Dir)).Call();
-			repo.Checkout().SetCreateBranch(true).SetUpstreamMode(CreateBranchCommand.SetupUpstreamMode.SET_UPSTREAM).SetName("origin/" + Branch).Call();
-			repo.SubmoduleInit().Call();
-			repo.SubmoduleUpdate().Call();
+			Method = method;
+			Permission = permission;
 		}
 	}
+	#pragma warning restore 3015
 }